refactor: re-sort
