+++ base/src/main/java/org/aya/tyck/ExprTycker.java
@@ -3,25 +3,27 @@
 package org.aya.tyck;
 
 import kala.collection.SeqView;
+import kala.collection.immutable.ImmutableMap;
 import kala.collection.immutable.ImmutableSeq;
 import kala.collection.mutable.MutableList;
 import kala.tuple.Tuple;
 import kala.tuple.Tuple2;
 import kala.tuple.Tuple3;
 import org.aya.concrete.Expr;
+import org.aya.concrete.stmt.decl.ClassDecl;
 import org.aya.concrete.stmt.decl.Decl;
+import org.aya.concrete.stmt.decl.TeleDecl;
 import org.aya.core.UntypedParam;
-import org.aya.core.def.DataDef;
-import org.aya.core.def.Def;
-import org.aya.core.def.MemberDef;
-import org.aya.core.def.PrimDef;
+import org.aya.core.def.*;
 import org.aya.core.repr.AyaShape;
+import org.aya.core.repr.CodeShape;
 import org.aya.core.term.*;
 import org.aya.core.visitor.AyaRestrSimplifier;
 import org.aya.core.visitor.Subst;
 import org.aya.core.visitor.Zonker;
 import org.aya.generic.Constants;
 import org.aya.tyck.tycker.UnifiedTycker;
+import org.aya.generic.SortKind;
 import org.aya.util.error.InternalException;
 import org.aya.guest0x0.cubical.CofThy;
 import org.aya.guest0x0.cubical.Partial;
@@ -48,6 +50,175 @@ import java.util.function.Function;
 public final class ExprTycker extends UnifiedTycker {
   public final @NotNull AyaShape.Factory shapeFactory;
 
+  public ExprTycker(@NotNull PrimDef.Factory primFactory, @NotNull AyaShape.Factory shapeFactory, @NotNull Reporter reporter, Trace.@Nullable Builder traceBuilder) {
+    super(reporter, traceBuilder, new TyckState(primFactory));
+    this.shapeFactory = shapeFactory;
+  }
+
+  private @NotNull Restr.Cond<Term> condition(@NotNull Restr.Cond<Expr> c) {
+    // forall i. (c_i is valid)
+    return new Restr.Cond<>(inherit(c.inst(), IntervalTerm.INSTANCE).wellTyped(), c.isOne());
+    // ^ note: `inst` may be ErrorTerm!
+  }
+
+  public @NotNull Partial<Term> elaboratePartial(@NotNull Expr.PartEl partial, @NotNull Term type) {
+    var s = new ClauseTyckState();
+    var sides = partial.clauses().flatMap(sys -> clause(sys.component1(), sys.component2(), type, s));
+    confluence(sides, partial, type);
+    if (s.isConstantFalse) return new Partial.Split<>(ImmutableSeq.empty());
+    if (s.truthValue != null) return new Partial.Const<>(s.truthValue);
+    return new Partial.Split<>(sides);
+  }
+
+  private static class ClauseTyckState {
+    public boolean isConstantFalse = false;
+    public @Nullable Term truthValue;
+  }
+
+  private @NotNull SeqView<Restr.Side<Term>> clause(@NotNull Expr lhs, @NotNull Expr rhs, @NotNull Term rhsType, @NotNull ClauseTyckState clauseState) {
+    return switch (AyaRestrSimplifier.INSTANCE.isOne(whnf(inherit(lhs, IntervalTerm.INSTANCE).wellTyped()))) {
+      case Restr.Disj<Term> restr -> {
+        var list = MutableList.<Restr.Side<Term>>create();
+        for (var cof : restr.orz()) {
+          var u = CofThy.vdash(cof, new Subst(), subst -> inherit(rhs, whnf(rhsType.subst(subst))).wellTyped());
+          if (u.isDefined()) {
+            if (u.get() == null) {
+              // ^ some `inst` in `cofib.ands()` are ErrorTerms, or we have bugs.
+              // Q: report error again?
+              yield SeqView.empty();
+            } else {
+              list.append(new Restr.Side<>(cof, u.get()));
+            }
+          }
+        }
+        yield list.view();
+      }
+      case Restr.Const<Term> c -> {
+        if (c.isOne()) clauseState.truthValue = inherit(rhs, rhsType).wellTyped();
+        else clauseState.isConstantFalse = true;
+        yield SeqView.empty();
+      }
+    };
+  }
+
+  private static final class NotPi extends Exception {
+    private final @NotNull Term what;
+
+    public NotPi(@NotNull Term what) {
+      this.what = what;
+    }
+  }
+
+  private Tuple2<PiTerm, @Nullable PathTerm>
+  ensurePiOrPath(@NotNull Term term) throws NotPi {
+    term = whnf(term);
+    if (term instanceof PiTerm pi) return Tuple.of(pi, null);
+    if (term instanceof PathTerm cube)
+      return Tuple.of(cube.computePi(), cube);
+    else throw new NotPi(term);
+  }
+
+  private static boolean needImplicitParamIns(@NotNull Expr expr) {
+    return expr instanceof Expr.Lambda ex && ex.param().explicit() || !(expr instanceof Expr.Lambda);
+  }
+
+  /// region Primary Functions
+
+  private @NotNull Result doInherit(@NotNull Expr expr, @NotNull Term term) {
+    return switch (expr) {
+      case Expr.Tuple(var pos, var it) -> {
+        var typeWHNF = whnf(term);
+        if (typeWHNF instanceof MetaTerm hole) yield inheritFallbackUnify(hole, synthesize(expr), expr);
+        if (!(typeWHNF instanceof SigmaTerm sigma))
+          yield fail(expr, term, BadTypeError.sigmaCon(state, expr, typeWHNF));
+        var resultTuple = sigma.check(it, (e, t) -> inherit(e, t).wellTyped());
+        if (resultTuple == null)
+          yield fail(expr, term, new TupleError.ElemMismatchError(pos, sigma.params().size(), it.size()));
+
+        yield new Result.Default(resultTuple, term);
+      }
+      case Expr.Hole hole -> {
+        // TODO[ice]: deal with unit type
+        var freshHole = ctx.freshHole(term, Constants.randomName(hole), hole.sourcePos());
+        if (hole.explicit()) reporter.report(new Goal(state, freshHole.component1(), hole.accessibleLocal().get()));
+        yield new Result.Default(freshHole.component2(), term);
+      }
+      case Expr.Lambda lam -> {
+        if (term instanceof MetaTerm) {
+          if (lam.param().type() instanceof Expr.Hole)
+            unifyTy(term, generatePi(lam), lam.param().sourcePos());
+          else yield inheritFallbackUnify(term, synthesize(lam), lam);
+        }
+        yield switch (whnf(term)) {
+          case PiTerm dt -> {
+            var param = lam.param();
+            if (param.explicit() != dt.param().explicit()) {
+              yield fail(lam, dt, new LicitError.LicitMismatch(lam, dt));
+            }
+            var var = param.ref();
+            var lamParam = param.type();
+            var type = dt.param().type();
+            var result = ty(lamParam);
+            if (unifyTyReported(result, type, lamParam))
+              type = result;
+            else yield error(lam, dt);
+            addWithTerm(param, type);
+            var resultParam = new Term.Param(var, type, param.explicit());
+            var body = dt.substBody(resultParam.toTerm());
+            yield ctx.with(resultParam, () -> {
+              var rec = check(lam.body(), body).wellTyped();
+              return new Result.Default(new LamTerm(LamTerm.param(resultParam), rec), dt);
+            });
+          }
+          // Path lambda!
+          case PathTerm path -> checkBoundaries(expr, path, new Subst(),
+            inherit(expr, path.computePi()).wellTyped());
+          default -> fail(lam, term, new BadExprError(lam, term));
+        };
+      }
+      case Expr.LitInt(var pos, var end) -> {
+        var ty = whnf(term);
+        if (ty instanceof IntervalTerm) {
+          if (end == 0 || end == 1) yield new Result.Default(end == 0 ? FormulaTerm.LEFT : FormulaTerm.RIGHT, ty);
+          else yield fail(expr, new PrimError.BadInterval(pos, end));
+        }
+        yield inheritFallbackUnify(term, synthesize(expr), expr);
+      }
+      case Expr.PartEl el -> {
+        if (!(whnf(term) instanceof PartialTyTerm ty)) yield fail(el, term, BadTypeError.partTy(state, el, term));
+        var cofTy = ty.restr();
+        var rhsType = ty.type();
+        var partial = elaboratePartial(el, rhsType);
+        var face = partial.restr();
+        if (!PartialTerm.impliesCof(cofTy, face, state))
+          yield fail(el, new CubicalError.FaceMismatch(el, face, cofTy));
+        yield new Result.Default(new PartialTerm(partial, rhsType), ty);
+      }
+      case Expr.Match match -> {
+        var discriminant = match.discriminant().map(this::synthesize);
+        var sig = new Def.Signature<>(discriminant.map(r -> new Term.Param(new LocalVar("_"), r.type(), true)), term);
+        var result = ClauseTycker.elabClausesClassified(this, match.clauses(), sig, match.sourcePos());
+        yield new Result.Default(new MatchTerm(discriminant.map(Result::wellTyped), result.matchings()), term);
+      }
+      case Expr.Let let -> checkLet(let, (body) -> check(body, term));
+      default -> inheritFallbackUnify(term, synthesize(expr), expr);
+    };
+  }
+
+  public @NotNull Result inherit(@NotNull Expr expr, @NotNull Term type) {
+    return traced(() -> new Trace.ExprT(expr, type.freezeHoles(state)), expr, e -> {
+      if (type instanceof PiTerm pi && !pi.param().explicit() && needImplicitParamIns(e)) {
+        var implicitParam = new Term.Param(new LocalVar(Constants.ANONYMOUS_PREFIX), pi.param().type(), false);
+        var body = ctx.with(implicitParam, () -> inherit(e, pi.substBody(implicitParam.toTerm()))).wellTyped();
+        return new Result.Default(new LamTerm(LamTerm.param(implicitParam), body), pi);
+      } else return doInherit(e, type);
+    });
+  }
+
+  public @NotNull Result check(@NotNull Expr expr, @NotNull Term type) {
+    return inherit(expr, type);
+
+  }
   private @NotNull Result doSynthesize(@NotNull Expr expr) {
     return switch (expr) {
       case Expr.Lambda lam when lam.param().type() instanceof Expr.Hole -> inherit(lam, generatePi(lam));
@@ -272,180 +443,8 @@ public final class ExprTycker extends UnifiedTycker {
     };
   }
 
-  /**
-   * tyck a let expr with the given checker
-   *
-   * @param checker check the type of the body of {@param let}
-   */
-  private @NotNull Result checkLet(@NotNull Expr.Let let, @NotNull Function<Expr, Result> checker) {
-    // pushing telescopes into lambda params, for example:
-    // `let f (x : A) : B x` is desugared to `let f : Pi (x : A) -> B x`
-    var letBind = let.bind();
-    var typeExpr = Expr.buildPi(letBind.sourcePos(),
-      letBind.telescope().view(), letBind.result());
-    // as well as the body of the binding, for example:
-    // `let f x := g` is desugared to `let f := \x => g`
-    var definedAsExpr = Expr.buildLam(letBind.sourcePos(),
-      letBind.telescope().view(), letBind.definedAs());
-
-    // Now everything is in the form of `let f : G := g in h`
-
-    // See the TeleDecl.FnDecl case of StmtTycker#tyckHeader
-    var type = ty(typeExpr).freezeHoles(state);
-    var definedAsResult = inherit(definedAsExpr, type);
-    var nameAndType = new Term.Param(let.bind().bindName(), definedAsResult.type(), true);
-
-    return subscoped(() -> {
-      definitionEqualities.addDirectly(nameAndType.ref(), definedAsResult.wellTyped(), definedAsResult.type());
-      return checker.apply(let.body());
-    });
-  }
-
-  public @NotNull Restr<Term> restr(@NotNull Restr<Expr> restr) {
-    return restr.mapCond(this::condition);
-  }
-
-  private @NotNull Restr.Cond<Term> condition(@NotNull Restr.Cond<Expr> c) {
-    // forall i. (c_i is valid)
-    return new Restr.Cond<>(inherit(c.inst(), IntervalTerm.INSTANCE).wellTyped(), c.isOne());
-    // ^ note: `inst` may be ErrorTerm!
-  }
-
-  public @NotNull Partial<Term> elaboratePartial(@NotNull Expr.PartEl partial, @NotNull Term type) {
-    var s = new ClauseTyckState();
-    var sides = partial.clauses().flatMap(sys -> clause(sys.component1(), sys.component2(), type, s));
-    confluence(sides, partial, type);
-    if (s.isConstantFalse) return new Partial.Split<>(ImmutableSeq.empty());
-    if (s.truthValue != null) return new Partial.Const<>(s.truthValue);
-    return new Partial.Split<>(sides);
-  }
-
-  private static class ClauseTyckState {
-    public boolean isConstantFalse = false;
-    public @Nullable Term truthValue;
-  }
-
-  private @NotNull SeqView<Restr.Side<Term>> clause(@NotNull Expr lhs, @NotNull Expr rhs, @NotNull Term rhsType, @NotNull ClauseTyckState clauseState) {
-    return switch (AyaRestrSimplifier.INSTANCE.isOne(whnf(inherit(lhs, IntervalTerm.INSTANCE).wellTyped()))) {
-      case Restr.Disj<Term> restr -> {
-        var list = MutableList.<Restr.Side<Term>>create();
-        for (var cof : restr.orz()) {
-          var u = CofThy.vdash(cof, new Subst(), subst -> inherit(rhs, whnf(rhsType.subst(subst))).wellTyped());
-          if (u.isDefined()) {
-            if (u.get() == null) {
-              // ^ some `inst` in `cofib.ands()` are ErrorTerms, or we have bugs.
-              // Q: report error again?
-              yield SeqView.empty();
-            } else {
-              list.append(new Restr.Side<>(cof, u.get()));
-            }
-          }
-        }
-        yield list.view();
-      }
-      case Restr.Const<Term> c -> {
-        if (c.isOne()) clauseState.truthValue = inherit(rhs, rhsType).wellTyped();
-        else clauseState.isConstantFalse = true;
-        yield SeqView.empty();
-      }
-    };
-  }
-
-  private static final class NotPi extends Exception {
-    private final @NotNull Term what;
-
-    public NotPi(@NotNull Term what) {
-      this.what = what;
-    }
-  }
-
-  private Tuple2<PiTerm, @Nullable PathTerm>
-  ensurePiOrPath(@NotNull Term term) throws NotPi {
-    term = whnf(term);
-    if (term instanceof PiTerm pi) return Tuple.of(pi, null);
-    if (term instanceof PathTerm cube)
-      return Tuple.of(cube.computePi(), cube);
-    else throw new NotPi(term);
-  }
-
-  private @NotNull Result doInherit(@NotNull Expr expr, @NotNull Term term) {
-    return switch (expr) {
-      case Expr.Tuple(var pos, var it) -> {
-        var typeWHNF = whnf(term);
-        if (typeWHNF instanceof MetaTerm hole) yield inheritFallbackUnify(hole, synthesize(expr), expr);
-        if (!(typeWHNF instanceof SigmaTerm sigma))
-          yield fail(expr, term, BadTypeError.sigmaCon(state, expr, typeWHNF));
-        var resultTuple = sigma.check(it, (e, t) -> inherit(e, t).wellTyped());
-        if (resultTuple == null)
-          yield fail(expr, term, new TupleError.ElemMismatchError(pos, sigma.params().size(), it.size()));
-        yield new Result.Default(resultTuple, term);
-      }
-      case Expr.Hole hole -> {
-        // TODO[ice]: deal with unit type
-        var freshHole = ctx.freshHole(term, Constants.randomName(hole), hole.sourcePos());
-        if (hole.explicit()) reporter.report(new Goal(state, freshHole.component1(), hole.accessibleLocal().get()));
-        yield new Result.Default(freshHole.component2(), term);
-      }
-      case Expr.Lambda lam -> {
-        if (term instanceof MetaTerm) {
-          if (lam.param().type() instanceof Expr.Hole)
-            unifyTy(term, generatePi(lam), lam.param().sourcePos());
-          else yield inheritFallbackUnify(term, synthesize(lam), lam);
-        }
-        yield switch (whnf(term)) {
-          case PiTerm dt -> {
-            var param = lam.param();
-            if (param.explicit() != dt.param().explicit()) {
-              yield fail(lam, dt, new LicitError.LicitMismatch(lam, dt));
-            }
-            var var = param.ref();
-            var lamParam = param.type();
-            var type = dt.param().type();
-            var result = ty(lamParam);
-            if (unifyTyReported(result, type, lamParam))
-              type = result;
-            else yield error(lam, dt);
-            addWithTerm(param, type);
-            var resultParam = new Term.Param(var, type, param.explicit());
-            var body = dt.substBody(resultParam.toTerm());
-            yield ctx.with(resultParam, () -> {
-              var rec = check(lam.body(), body).wellTyped();
-              return new Result.Default(new LamTerm(LamTerm.param(resultParam), rec), dt);
-            });
-          }
-          // Path lambda!
-          case PathTerm path -> checkBoundaries(expr, path, new Subst(),
-            inherit(expr, path.computePi()).wellTyped());
-          default -> fail(lam, term, new BadExprError(lam, term));
-        };
-      }
-      case Expr.LitInt(var pos, var end) -> {
-        var ty = whnf(term);
-        if (ty instanceof IntervalTerm) {
-          if (end == 0 || end == 1) yield new Result.Default(end == 0 ? FormulaTerm.LEFT : FormulaTerm.RIGHT, ty);
-          else yield fail(expr, new PrimError.BadInterval(pos, end));
-        }
-        yield inheritFallbackUnify(term, synthesize(expr), expr);
-      }
-      case Expr.PartEl el -> {
-        if (!(whnf(term) instanceof PartialTyTerm ty)) yield fail(el, term, BadTypeError.partTy(state, el, term));
-        var cofTy = ty.restr();
-        var rhsType = ty.type();
-        var partial = elaboratePartial(el, rhsType);
-        var face = partial.restr();
-        if (!PartialTerm.impliesCof(cofTy, face, state))
-          yield fail(el, new CubicalError.FaceMismatch(el, face, cofTy));
-        yield new Result.Default(new PartialTerm(partial, rhsType), ty);
-      }
-      case Expr.Match match -> {
-        var discriminant = match.discriminant().map(this::synthesize);
-        var sig = new Def.Signature<>(discriminant.map(r -> new Term.Param(new LocalVar("_"), r.type(), true)), term);
-        var result = ClauseTycker.elabClausesClassified(this, match.clauses(), sig, match.sourcePos());
-        yield new Result.Default(new MatchTerm(discriminant.map(Result::wellTyped), result.matchings()), term);
-      }
-      case Expr.Let let -> checkLet(let, (body) -> check(body, term));
-      default -> inheritFallbackUnify(term, synthesize(expr), expr);
-    };
+  public @NotNull Result synthesize(@NotNull Expr expr) {
+    return traced(() -> new Trace.ExprT(expr, null), expr, this::doSynthesize);
   }
 
   private @NotNull Term doTy(@NotNull Expr expr) {
@@ -480,6 +479,16 @@ public final class ExprTycker extends UnifiedTycker {
     };
   }
 
+  public @NotNull Term ty(@NotNull Expr expr) {
+    return traced(() -> new Trace.ExprT(expr, null), () -> doTy(expr));
+  }
+
+  /// endregion Primary Functions
+
+  /// region Particular Type Checking
+
+  // Sort:
+
   public @NotNull Result.Sort sort(@NotNull Expr expr) {
     return new Result.Sort(sort(expr, ty(expr)));
   }
@@ -498,34 +507,73 @@ public final class ExprTycker extends UnifiedTycker {
     };
   }
 
-  public ExprTycker(@NotNull PrimDef.Factory primFactory, @NotNull AyaShape.Factory shapeFactory, @NotNull Reporter reporter, Trace.@Nullable Builder traceBuilder) {
-    super(reporter, traceBuilder, new TyckState(primFactory));
-    this.shapeFactory = shapeFactory;
-  }
+  // Let:
 
-  public @NotNull Result inherit(@NotNull Expr expr, @NotNull Term type) {
-    return traced(() -> new Trace.ExprT(expr, type.freezeHoles(state)), expr, e -> {
-      if (type instanceof PiTerm pi && !pi.param().explicit() && needImplicitParamIns(e)) {
-        var implicitParam = new Term.Param(new LocalVar(Constants.ANONYMOUS_PREFIX), pi.param().type(), false);
-        var body = ctx.with(implicitParam, () -> inherit(e, pi.substBody(implicitParam.toTerm()))).wellTyped();
-        return new Result.Default(new LamTerm(LamTerm.param(implicitParam), body), pi);
-      } else return doInherit(e, type);
+  /**
+   * tyck a let expr with the given checker
+   *
+   * @param checker check the type of the body of {@param let}
+   */
+  private @NotNull Result checkLet(@NotNull Expr.Let let, @NotNull Function<Expr, Result> checker) {
+    // pushing telescopes into lambda params, for example:
+    // `let f (x : A) : B x` is desugared to `let f : Pi (x : A) -> B x`
+    var letBind = let.bind();
+    var typeExpr = Expr.buildPi(letBind.sourcePos(),
+      letBind.telescope().view(), letBind.result());
+    // as well as the body of the binding, for example:
+    // `let f x := g` is desugared to `let f := \x => g`
+    var definedAsExpr = Expr.buildLam(letBind.sourcePos(),
+      letBind.telescope().view(), letBind.definedAs());
+
+    // Now everything is in the form of `let f : G := g in h`
+
+    // See the TeleDecl.FnDecl case of StmtTycker#tyckHeader
+    var type = ty(typeExpr).freezeHoles(state);
+    var definedAsResult = inherit(definedAsExpr, type);
+    var nameAndType = new Term.Param(let.bind().bindName(), definedAsResult.type(), true);
+
+    return subscoped(() -> {
+      definitionEqualities.addDirectly(nameAndType.ref(), definedAsResult.wellTyped(), definedAsResult.type());
+      return checker.apply(let.body());
     });
   }
 
-  public @NotNull Result synthesize(@NotNull Expr expr) {
-    return traced(() -> new Trace.ExprT(expr, null), expr, this::doSynthesize);
+  // Restr:
+
+  public @NotNull Restr<Term> restr(@NotNull Restr<Expr> restr) {
+    return restr.mapCond(this::condition);
   }
 
-  public @NotNull Term ty(@NotNull Expr expr) {
-    return traced(() -> new Trace.ExprT(expr, null), () -> doTy(expr));
+  /// endregion Particular Type Checking
+
+  /// region Helpful Utils
+
+  @SuppressWarnings("unchecked") private @NotNull Result inferRef(@NotNull DefVar<?, ?> var) {
+    if (var.core instanceof FnDef || var.concrete instanceof TeleDecl.FnDecl) {
+      return defCall((DefVar<FnDef, TeleDecl.FnDecl>) var, FnCall::new);
+    } else if (var.core instanceof PrimDef) {
+      return defCall((DefVar<PrimDef, TeleDecl.PrimDecl>) var, PrimCall::new);
+    } else if (var.core instanceof DataDef || var.concrete instanceof TeleDecl.DataDecl) {
+      return defCall((DefVar<DataDef, TeleDecl.DataDecl>) var, DataCall::new);
+    } else if (var.core instanceof ClassDef || var.concrete instanceof ClassDecl) {
+      var classCall = new ClassCall((DefVar<ClassDef, ClassDecl>) var, 0, ImmutableMap.empty());
+      return new Result.Default(classCall, new SortTerm(SortKind.Type, 0)); // TODO[class]: type of classCall
+    } else if (var.core instanceof CtorDef || var.concrete instanceof TeleDecl.DataDecl.DataCtor) {
+      var conVar = (DefVar<CtorDef, TeleDecl.DataDecl.DataCtor>) var;
+      var tele = Def.defTele(conVar);
+      var type = PiTerm.make(tele, Def.defResult(conVar)).rename();
+      var telescopes = new DataDef.CtorTelescopes(conVar.core);
+      return new Result.Default(telescopes.toConCall(conVar, 0), type);
+    } else if (var.core instanceof MemberDef || var.concrete instanceof TeleDecl.ClassMember) {
+      // the code runs to here because we are checking a StructField within a StructDecl
+      // TODO[class]: this needs to be refactored to make use of instance resolution
+      var field = (DefVar<MemberDef, TeleDecl.ClassMember>) var;
+      return new Result.Default(new RefTerm.Field(field), Def.defType(field));
+    } else {
+      final var msg = "Def var `" + var.name() + "` has core `" + var.core + "` which we don't know.";
+      throw new InternalException(msg);
+    }
   }
 
-  private static boolean needImplicitParamIns(@NotNull Expr expr) {
-    return expr instanceof Expr.Lambda ex && ex.param().explicit() || !(expr instanceof Expr.Lambda);
-  }
-
-  public @NotNull Result check(@NotNull Expr expr, @NotNull Term type) {
-    return inherit(expr, type);
-  }
+  /// endregion Helpful Utils
 }
